requirements.txt
langchain
langchain-community
langchain-huggingface
streamlit
python-dotenv